// src/app/credit-request/models/fileRepositoryOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { FileSetOutputModel } from './fileSetOutputModel';

/**
 * File Sets in the repository
 */
export interface FileRepositoryOutputModel {
    /**
     * Id of the file repository
     */
    fileRepositoryId?: string;
    /**
     * Details text
     */
    details?: string;
    /**
     * List of sets of documents in the repository
     */
    fileSets?: Array<FileSetOutputModel>;
}